navdp_api/requirements.txt
torch==2.2.2
diffusers==0.33.1
Flask==3.1.1
gradio==5.31.0
gradio_imageslider==0.0.20
h5py==3.13.0
matplotlib==3.8.4
numpy==1.26.0
open3d==0.19.0
opencv_python==4.9.0.80
Pillow==11.2.1
Requests==2.32.3
torchvision==0.17.2
imageio[ffmpeg]==2.31.1